Package requirements
With this commit our package now also defines its requirements to be
installed alond with it. This is nice and makes the package usable as
one would expect it.
With this change, I could tweak the imports to be correct. Doing so I
noticed that I would need a hell lot of work to get the custom events
that are projectQ specific working. But as I wanted to remove them
anyway, I just removed them now along with their tests. Those events
need to be specified by the package using this one.